Return flight actually taken into account
Previously, we would select the date of the return flight even if the radio button for one way flight was selected. This is now fixed. Also, a better seat taken detection system is implemented (a hash map is used rather than checking the foreground color). Some error messages are changed